refactor(harness-core/config): extract tests via #[path]
Phase 0 PR-6 of the architecture decoupling roadmap (#1082, #1083,
#1084, #1086, #1088). Cleanest mechanical victory in the series:
extraction brings config.rs fully under the U-16 800-line ceiling.

Net effect:
  harness-core/src/config.rs:        1095 -> 214 lines (-881)
  harness-core/src/config_tests.rs:   NEW   881 lines

Same pattern: `#[cfg(test)] #[path = "config_tests.rs"] mod tests;`
with the body lifted verbatim and dedented one level.

Verification:
- cargo test -p harness-core --lib -- config:
    143/143 passed (covers config::tests::* plus cross-module
    consumers db_pg::tests::*, prompts::*, etc.)
- RUSTFLAGS=-Dwarnings cargo check --workspace --all-targets: clean
- cargo fmt --all -- --check: clean

config.rs is now 214 production lines, well under U-16. This is the
first Phase 0 PR where extraction completes the file-size victory
in one step (worker.rs / submission.rs / workspace.rs / execution.rs
/ spawn.rs all still need Phase 1 cohesive splits to reach < 800).
